+++ Utils.lua
@@ -522,6 +522,7 @@ KART.Defaults = {
     inviteKeywords = "inv;+;invite",
     inviteViaGuildChat = false,
     inviteChannels = { WHISPER = true, BN = true, GUILD = false, OFFICER = false },
+    inviteGuildRanks = {},
     promoteNames = "",
     showRaidleadBar = false,
     autoModuleEnabled = true,