+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- **Invite online guild members by rank** from the Automation tab. Select one or more ranks, then confirm the invite.
 
 ## [4.3.0] - 2026-09-04
 ### Added